+++ libs/pixi-component/engine/scene.ts
@@ -3,7 +3,7 @@ import Message from './message';
 import Component from './component';
 import * as PIXI from 'pixi.js';
 import { Messages, AttributeChangeMessage, StateChangeMessage, FlagChangeMessage, TagChangeMessage } from './constants';
-import { GameObject, Container } from './game-object';
+import { Container } from './game-object';
 import { LookupMap } from '../utils/lookup-map';
 import DebugComponent from '../components/debug-component';
 import { QueryCondition, queryConditionCheck } from '../utils/query-condition';
@@ -50,29 +50,31 @@ export default class Scene {
   app: PIXI.Application;
   name: string;
   // PIXI stage object
-  stage: GameObject = null;
+  stage: Container = null;
   // collection of actions that should be invoked with a delay
   private pendingInvocations: Invocation[];
   // message action keys and all subscribers that listens to all these actions
   private subscribers: LookupMap<string, Component> ;
   // game objects mapped by their flags
-  private gameObjectFlags: LookupMap<number, GameObject>;
+  private gameObjectFlags: LookupMap<number, Container>;
   // game objects mapped by their state
-  private gameObjectStates: LookupMap<number, GameObject>;
+  private gameObjectStates: LookupMap<number, Container>;
   // game objects mapped by their tags
-  private gameObjectTags: LookupMap<string, GameObject>;
+  private gameObjectTags: LookupMap<string, Container>;
   // game objects mapped by their names
-  private gameObjectNames: LookupMap<string, GameObject>;
+  private gameObjectNames: LookupMap<string, Container>;
   // collection of ALL game objects, mapped by their ids
-  private gameObjects: Map<number, GameObject>;
+  private gameObjects: Map<number, Container>;
+  // indicator if the scene is just being updated
+  private updating: boolean;
   // COMPONENT_ADDED event will not be fired if a new object is being added to the scene
   private componentNotifyDisabled = false;
-  protected _currentDelta: number;
-  protected _currentAbsolute: number;
+  private _currentDelta: number;
+  private _currentAbsolute: number;
 
-  protected config: SceneConfig;
+  private config: SceneConfig;
   // indicator that will be reset upon first update
-  protected sceneCleared: boolean;
+  private sceneCleared: boolean;
 
   constructor(name: string, app: PIXI.Application, config?: SceneConfig) {
     this.name = name;
@@ -150,7 +152,7 @@ export default class Scene {
   /**
    * Gets object by its id
    */
-  getObjectById(id: number): GameObject {
+  getObjectById(id: number): Container {
     if(this.gameObjects.has(id)) {
       return this.gameObjects.get(id);
     }
@@ -160,8 +162,8 @@ export default class Scene {
   /**
    * Finds all objects that meet specific condition
    */
-  findObjectsByQuery(query: QueryCondition): Array<GameObject> {
-    let result: GameObject[] = [];
+  findObjectsByQuery(query: QueryCondition): Array<Container> {
+    let result: Container[] = [];
     // linear complexity
     for(let [,object] of this.gameObjects) {
       if(queryConditionCheck(object, query)) {
@@ -174,7 +176,7 @@ export default class Scene {
   /**
    * Finds all game objects by their name
    */
-  findObjectsByName(name: string): Array<GameObject> {
+  findObjectsByName(name: string): Array<Container> {
     if(!this.config.namesSearchEnabled) {
       throw new Error('Searching by name is not enabled. See SceneConfig');
     }
@@ -184,7 +186,7 @@ export default class Scene {
   /**
    * Finds a first object with a given name
    */
-  findObjectByName(name: string): GameObject {
+  findObjectByName(name: string): Container {
     if(!this.config.namesSearchEnabled) {
       throw new Error('Searching by name is not enabled. See SceneConfig');
     }
@@ -194,7 +196,7 @@ export default class Scene {
   /**
    * Finds all game objects by their tag
    */
-  findObjectsByTag(tag: string): Array<GameObject> {
+  findObjectsByTag(tag: string): Array<Container> {
     if(!this.config.tagsSearchEnabled) {
       throw new Error('Searching by tag is not enabled. See SceneConfig');
     }
@@ -204,7 +206,7 @@ export default class Scene {
   /**
    * Finds a first object with a given tag
    */
-  findObjectByTag(tag: string): GameObject {
+  findObjectByTag(tag: string): Container {
     if(!this.config.tagsSearchEnabled) {
       throw new Error('Searching by tag is not enabled. See SceneConfig');
     }
@@ -214,7 +216,7 @@ export default class Scene {
   /**
    * Finds all game objects by their flag
    */
-  findObjectsByFlag(flag: number): Array<GameObject> {
+  findObjectsByFlag(flag: number): Array<Container> {
     if(!this.config.flagsSearchEnabled) {
       throw new Error('Searching by flags is not enabled. See SceneConfig');
     }
@@ -224,7 +226,7 @@ export default class Scene {
   /**
    * Finds a first object with a given flag
    */
-  findObjectByFlag(flag: number): GameObject {
+  findObjectByFlag(flag: number): Container {
     if(!this.config.flagsSearchEnabled) {
       throw new Error('Searching by flags is not enabled. See SceneConfig');
     }
@@ -234,7 +236,7 @@ export default class Scene {
   /**
    * Finds all game objects by their state
    */
-  findObjectsByState(state: number): Array<GameObject> {
+  findObjectsByState(state: number): Array<Container> {
     if(!this.config.statesSearchEnabled) {
       throw new Error('Searching by states is not enabled. See SceneConfig');
     }
@@ -244,7 +246,7 @@ export default class Scene {
   /**
    * Finds a first object with a given state
    */
-  findObjectByState(state: number): GameObject {
+  findObjectByState(state: number): Container {
     if(!this.config.statesSearchEnabled) {
       throw new Error('Searching by states is not enabled. See SceneConfig');
     }
@@ -271,6 +273,10 @@ export default class Scene {
    * Removes all objects from scene
    */
   clearScene(newConfig?: SceneConfig) {
+    if(this.updating) {
+      throw new Error('Scene can\'t be cleared during update. Use invokeWithDelay() instead!');
+    }
+
     this.sendMessage(new Message(Messages.SCENE_CLEAR, null, null, this.name));
     // call the finalization function of all components
     for (let [, gameObj] of this.gameObjects) {
@@ -349,8 +355,10 @@ export default class Scene {
     this._currentDelta = delta;
     this._currentAbsolute = absolute;
 
+    this.updating = true;
     // update root object and all other objects recursively
     this.stage._proxy.update(delta, absolute);
+    this.updating = false;
 
     // execute pending invocations
     let i = this.pendingInvocations.length;
@@ -383,7 +391,7 @@ export default class Scene {
   }
 
   _onObjectAdded(obj: GameObjectProxy) {
-    let pixiObj = obj.cmpObj;
+    let pixiObj = obj.pixiObj;
     // fill all collections
     if(this.config.namesSearchEnabled) {
       this.gameObjectNames.insert(pixiObj.name, pixiObj);
@@ -401,7 +409,7 @@ export default class Scene {
       this.gameObjectStates.insert(pixiObj.stateId, pixiObj);
     }
 
-    this.gameObjects.set(obj.id, obj.cmpObj);
+    this.gameObjects.set(obj.id, obj.pixiObj);
 
     // assign scene
     obj.scene = this;
@@ -414,7 +422,7 @@ export default class Scene {
   }
 
   _onObjectRemoved(obj: GameObjectProxy) {
-    let gameObj = obj.cmpObj;
+    let gameObj = obj.pixiObj;
 
     if(this.config.namesSearchEnabled) {
       this.gameObjectNames.remove(gameObj.name, gameObj);
@@ -441,23 +449,23 @@ export default class Scene {
   _onComponentAdded(component: Component, obj: GameObjectProxy) {
     component.scene = this;
     if(!this.componentNotifyDisabled) {
-      this.sendMessage(new Message(Messages.COMPONENT_ADDED, component, obj.cmpObj));
+      this.sendMessage(new Message(Messages.COMPONENT_ADDED, component, obj.pixiObj));
     }
   }
 
   _onComponentRemoved(component: Component, obj: GameObjectProxy) {
     this.subscribers.removeItem(component);
-    this.sendMessage(new Message(Messages.COMPONENT_REMOVED, component, obj.cmpObj));
+    this.sendMessage(new Message(Messages.COMPONENT_REMOVED, component, obj.pixiObj));
   }
 
   _onStateChanged(previous: number, current: number, obj: GameObjectProxy) {
     if(this.config.statesSearchEnabled) {
-      this.gameObjectStates.remove(previous, obj.cmpObj);
-      this.gameObjectStates.insert(current, obj.cmpObj);
+      this.gameObjectStates.remove(previous, obj.pixiObj);
+      this.gameObjectStates.insert(current, obj.pixiObj);
     }
     if(this.config.notifyStateChanges) {
       let data: StateChangeMessage = { previous, current };
-      this.sendMessage(new Message(Messages.STATE_CHANGED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.STATE_CHANGED, null, obj.pixiObj, data));
     }
   }
 
@@ -469,7 +477,7 @@ export default class Scene {
         previousValue: null,
         currentValue: value
       };
-      this.sendMessage(new Message(Messages.ATTRIBUTE_ADDED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.ATTRIBUTE_ADDED, null, obj.pixiObj, data));
     }
   }
 
@@ -481,7 +489,7 @@ export default class Scene {
         previousValue: previousValue,
         currentValue: currentValue
       };
-      this.sendMessage(new Message(Messages.ATTRIBUTE_CHANGED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.ATTRIBUTE_CHANGED, null, obj.pixiObj, data));
     }
   }
 
@@ -493,41 +501,41 @@ export default class Scene {
         previousValue: value,
         currentValue: null
       };
-      this.sendMessage(new Message(Messages.ATTRIBUTE_REMOVED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.ATTRIBUTE_REMOVED, null, obj.pixiObj, data));
     }
   }
 
   _onFlagChanged(flag: number, set: boolean, obj: GameObjectProxy) {
     if(this.config.flagsSearchEnabled) {
       if(set) {
-        this.gameObjectFlags.insert(flag, obj.cmpObj);
+        this.gameObjectFlags.insert(flag, obj.pixiObj);
       } else {
-        this.gameObjectFlags.remove(flag, obj.cmpObj);
+        this.gameObjectFlags.remove(flag, obj.pixiObj);
       }
     }
     if(this.config.notifyFlagChanges) {
       let data: FlagChangeMessage = { flag, isSet: set };
-      this.sendMessage(new Message(Messages.FLAG_CHANGED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.FLAG_CHANGED, null, obj.pixiObj, data));
     }
   }
 
   _onTagAdded(tag: string, obj: GameObjectProxy) {
     if(this.config.tagsSearchEnabled) {
-      this.gameObjectTags.insert(tag, obj.cmpObj);
+      this.gameObjectTags.insert(tag, obj.pixiObj);
     }
     if(this.config.notifyTagChanges) {
       let data: TagChangeMessage = { tag, type: Messages.TAG_ADDED };
-      this.sendMessage(new Message(Messages.TAG_ADDED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.TAG_ADDED, null, obj.pixiObj, data));
     }
   }
 
   _onTagRemoved(tag: string, obj: GameObjectProxy) {
     if(this.config.tagsSearchEnabled) {
-      this.gameObjectTags.remove(tag, obj.cmpObj);
+      this.gameObjectTags.remove(tag, obj.pixiObj);
     }
     if(this.config.notifyTagChanges) {
       let data: TagChangeMessage = { tag, type: Messages.TAG_REMOVED };
-      this.sendMessage(new Message(Messages.TAG_REMOVED, null, obj.cmpObj, data));
+      this.sendMessage(new Message(Messages.TAG_REMOVED, null, obj.pixiObj, data));
     }
   }