+++ libs/pixi-math/index.ts
@@ -1,0 +1,8 @@
+import Random from './math/random';
+import { MatterBodyOptions, MatterBody, MatterConstraintOptions, MatterConstraint } from './math/matter-components';
+import { QuadTree, QuadTreeItem } from './math/quad-tree';
+import { SteeringMath } from './math/steering';
+
+export {
+  Random, MatterBodyOptions, MatterBody, MatterConstraintOptions, MatterConstraint, QuadTree, QuadTreeItem, SteeringMath
+};