+++ src/labs/lab03/squares.ts
@@ -1,6 +1,32 @@
 
 import * as ECSA from '../../../libs/pixi-component';
 
+const FLAG_ROTATING = 1;
+
+class RotationAnim extends ECSA.Component {
+
+  onInit() {
+    this.subscribe('ROTATION_FINISHED');
+  }
+
+  onMessage(msg:  ECSA.Message) {
+    if(msg.action === 'ROTATION_FINISHED') {
+      this.owner.setFlag(FLAG_ROTATING);
+    }
+  }
+
+  onUpdate(delta: number, absolute: number) {
+    if(this.owner.hasFlag(FLAG_ROTATING)) {
+      this.owner.asContainer().rotation += delta * 0.004;
+      if(this.owner.asContainer().rotation >= 2*Math.PI) {
+        this.owner.asContainer().rotation = 0;
+        this.sendMessage('ROTATION_FINISHED');
+        this.owner.resetFlag(FLAG_ROTATING);
+      }
+    }
+  }
+}
+
 class Squares {
   engine: ECSA.GameLoop;
 
@@ -20,10 +46,13 @@ class Squares {
     square2.drawRect(0, 0, 200, 200);
     square2.endFill();
 
-    let builder = new ECSA.Builder(this.engine.scene).withParent(this.engine.scene.stage).anchor(0.5);
+    let builder = new ECSA.Builder(this.engine.scene)
+    .withComponent(() => new RotationAnim())
+    .withParent(this.engine.scene.stage)
+    .anchor(0.5);
 
     builder.relativePos(0.25, 0.5).buildInto(square1, false);
-    builder.relativePos(0.75, 0.5).buildInto(square2, false);
+    builder.relativePos(0.75, 0.5).withFlag(FLAG_ROTATING).buildInto(square2, false);
   }
 }